# Optimize Bisection API and Area Enumeration