--- a/TheSadRogue.Primitives/Area.cs
+++ b/TheSadRogue.Primitives/Area.cs
@@ -448,13 +448,13 @@ namespace SadRogue.Primitives
         /// Returns an enumerator that iterates through all positions in the Area, in the order they were added.
         /// </summary>
         /// <returns>An enumerator that iterates through all the positions in the Area, in the order they were added.</returns>
-        public IEnumerator<Point> GetEnumerator() => _positions.GetEnumerator();
+        public ReadOnlyAreaPostionsEnumerable GetEnumerator() => new ReadOnlyAreaPostionsEnumerable(this);
 
         /// <summary>
         /// Returns an enumerator that iterates through all positions in the Area, in the order they were added.
         /// </summary>
         /// <returns>An enumerator that iterates through all the positions in the Area, in the order they were added.</returns>
-        IEnumerator IEnumerable.GetEnumerator() => ((IEnumerable)_positions).GetEnumerator();
+        //IEnumerator IEnumerable.GetEnumerator() => ((IEnumerable)_positions).GetEnumerator();
 
         private void RecalculateBounds()
         {