--- a/TheSadRogue.Primitives/IReadOnlyArea.cs
+++ b/TheSadRogue.Primitives/IReadOnlyArea.cs
@@ -1,7 +1,43 @@
 ﻿using System.Collections.Generic;
+using System.Runtime.CompilerServices;
 
 namespace SadRogue.Primitives
 {
+    public struct ReadOnlyAreaPostionsEnumerable
+    {
+        private readonly IReadOnlyArea _area;
+
+        private readonly int _count;
+        private int _currentIdx;
+        public Point Current => _area[_currentIdx];
+
+        public ReadOnlyAreaPostionsEnumerable(IReadOnlyArea area)
+        {
+            _area = area;
+            _currentIdx = -1;
+            _count = area.Count;
+        }
+
+        public bool MoveNext()
+        {
+            _currentIdx++;
+            return _currentIdx < _count;
+        }
+
+        /// <summary>
+        /// Returns this enumerator.
+        /// </summary>
+        /// <returns>This enumerator.</returns>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public ReadOnlyAreaPostionsEnumerable GetEnumerator() => this;
+
+        public IEnumerable<Point> ToEnumerable()
+        {
+            int count = _area.Count;
+            for (int i = 0; i < count; i++)
+                yield return _area[i];
+        }
+    }
     /// <summary>
     /// Read-only interface for an arbitrary 2D area.
     /// </summary>