… TheSadRogue.Primitives/ReadOnlyAreaExtensions.cs
… TheSadRogue.Primitives/ReadOnlyAreaPositionsEnumerable.cs
… TheSadRogue.Primitives/Rectangle.cs
… TheSadRogue.Primitives/RectanglePositionsEnumerable.cs
… TheSadRogue.Primitives/SerializedTypes/Area.cs
… TheSadRogue.Primitives/SerializedTypes/Rectangle.cs
… TheSadRogue.Primitives/TheSadRogue.Primitives.csproj