# Changes
- `Bisect*` functions for rectangle now return a value type of two rectangles, instead of an enumerable of rectangles, where possible (closes #72 )
    - The value type returned from these functions supports deconstruction, converts implicitly to an equivalent value tuple, and has a `ToEnumerable` function which returns an enumerable of both rectangles.
- Manually expanded outer for loop in `BisectRecursive`
- Added a `FastEnumerator()` extension method to `IReadOnlyArea`, which provides a quick way to iterate over all points in an area
    - Allows generic `IReadOnlyArea` functions to take advantage of the speed benefits of iterating via index for list-based area implmentations
- Added optional parameter to `GetIntersection`, `GetUnion`, and `GetDifference` functions which allows the use of a custom point hasher for the resulting area

# Notes
Not much compatibility breakage here, just adding a `ToEnumerable` if a user really needs an enumerable from these functions/classes.  The performance implications are fairly straightforward.

## Rectangle Bisection
Here are the performance test results.  The `BisectHorizontallyNativeEnumerable` shows the performance of the original `BisectHorizontal` implementation, for reference (the vertical function was similar).  These tests show 3x faster in the new implementation.

|                             Method | Width | Height |           Mean |         Error |        StdDev |         Median |
|----------------------------------- |------ |------- |---------------:|--------------:|--------------:|---------------:|
|                             Bisect |    10 |     10 |      11.960 ns |     0.0129 ns |     0.0121 ns |      11.954 ns |
|                   BisectVertically |    10 |     10 |      10.737 ns |     0.0414 ns |     0.0367 ns |      10.716 ns |
|                 BisectHorizontally |    10 |     10 |       9.898 ns |     0.0323 ns |     0.0302 ns |       9.889 ns |
|                    BisectRecursive |    10 |     10 |     142.001 ns |     1.9671 ns |     1.6426 ns |     141.642 ns |
|                 BisectToEnumerable |    10 |     10 |      31.704 ns |     0.6614 ns |     1.7541 ns |      31.113 ns |
|       BisectVerticallyToEnumerable |    10 |     10 |      29.138 ns |     0.5555 ns |     0.7224 ns |      29.141 ns |
|     BisectHorizontallyToEnumerable |    10 |     10 |      31.083 ns |     0.6209 ns |     0.8905 ns |      30.842 ns |
| BisectHorizontallyNativeEnumerable |    10 |     10 |      26.888 ns |     0.4518 ns |     0.8487 ns |      26.648 ns |
|                             Bisect |    10 |     50 |      11.793 ns |     0.0711 ns |     0.0555 ns |      11.781 ns |
|                   BisectVertically |    10 |     50 |      10.284 ns |     0.0334 ns |     0.0372 ns |      10.284 ns |
|                 BisectHorizontally |    10 |     50 |      10.197 ns |     0.1890 ns |     0.1578 ns |      10.140 ns |
|                    BisectRecursive |    10 |     50 |   2,266.148 ns |    21.7679 ns |    18.1772 ns |   2,267.865 ns |
|                 BisectToEnumerable |    10 |     50 |      31.041 ns |     0.6340 ns |     0.5930 ns |      30.997 ns |
|       BisectVerticallyToEnumerable |    10 |     50 |      30.692 ns |     0.5786 ns |     0.4831 ns |      30.808 ns |
|     BisectHorizontallyToEnumerable |    10 |     50 |      30.677 ns |     0.6362 ns |     1.3831 ns |      30.247 ns |
| BisectHorizontallyNativeEnumerable |    10 |     50 |      26.754 ns |     0.4856 ns |     0.4542 ns |      26.711 ns |
|                             Bisect |    10 |    100 |      11.804 ns |     0.0967 ns |     0.0808 ns |      11.778 ns |
|                   BisectVertically |    10 |    100 |      10.846 ns |     0.0631 ns |     0.0492 ns |      10.851 ns |
|                 BisectHorizontally |    10 |    100 |      10.585 ns |     0.1858 ns |     0.1647 ns |      10.544 ns |
|                    BisectRecursive |    10 |    100 |   5,424.528 ns |   107.9690 ns |   154.8459 ns |   5,368.166 ns |
|                 BisectToEnumerable |    10 |    100 |      31.556 ns |     0.6530 ns |     1.3338 ns |      31.466 ns |
|       BisectVerticallyToEnumerable |    10 |    100 |      30.464 ns |     0.7921 ns |     2.2342 ns |      29.579 ns |
|     BisectHorizontallyToEnumerable |    10 |    100 |      28.284 ns |     0.4962 ns |     0.4399 ns |      28.291 ns |
| BisectHorizontallyNativeEnumerable |    10 |    100 |      26.791 ns |     0.5583 ns |     0.6646 ns |      26.755 ns |
|                             Bisect |    50 |     10 |      11.545 ns |     0.0845 ns |     0.0660 ns |      11.537 ns |
|                   BisectVertically |    50 |     10 |      10.745 ns |     0.0411 ns |     0.0364 ns |      10.754 ns |
|                 BisectHorizontally |    50 |     10 |      10.206 ns |     0.0555 ns |     0.1147 ns |      10.189 ns |
|                    BisectRecursive |    50 |     10 |   2,319.263 ns |    45.9436 ns |   104.6367 ns |   2,289.259 ns |
|                 BisectToEnumerable |    50 |     10 |      32.432 ns |     0.6637 ns |     1.5250 ns |      32.092 ns |
|       BisectVerticallyToEnumerable |    50 |     10 |      29.696 ns |     0.6256 ns |     1.1439 ns |      29.445 ns |
|     BisectHorizontallyToEnumerable |    50 |     10 |      29.636 ns |     0.6019 ns |     0.7612 ns |      29.497 ns |
| BisectHorizontallyNativeEnumerable |    50 |     10 |      27.019 ns |     0.5648 ns |     0.5800 ns |      26.838 ns |
|                             Bisect |    50 |     50 |      12.243 ns |     0.1497 ns |     0.1250 ns |      12.240 ns |
|                   BisectVertically |    50 |     50 |      11.420 ns |     0.5300 ns |     1.5627 ns |      10.449 ns |
|                 BisectHorizontally |    50 |     50 |      10.206 ns |     0.1606 ns |     0.2251 ns |      10.128 ns |
|                    BisectRecursive |    50 |     50 |  24,490.356 ns |   214.9100 ns |   179.4596 ns |  24,436.105 ns |
|                 BisectToEnumerable |    50 |     50 |      34.794 ns |     1.6003 ns |     4.6935 ns |      32.636 ns |
|       BisectVerticallyToEnumerable |    50 |     50 |      32.724 ns |     1.8708 ns |     5.4867 ns |      29.505 ns |
|     BisectHorizontallyToEnumerable |    50 |     50 |      29.063 ns |     0.4275 ns |     0.9908 ns |      28.699 ns |
| BisectHorizontallyNativeEnumerable |    50 |     50 |      29.996 ns |     1.4358 ns |     4.2109 ns |      28.911 ns |
|                             Bisect |    50 |    100 |      12.181 ns |     0.1153 ns |     0.0962 ns |      12.163 ns |
|                   BisectVertically |    50 |    100 |      10.162 ns |     0.0737 ns |     0.0615 ns |      10.145 ns |
|                 BisectHorizontally |    50 |    100 |       9.914 ns |     0.0464 ns |     0.0387 ns |       9.911 ns |
|                    BisectRecursive |    50 |    100 |  52,368.069 ns | 1,007.3419 ns | 1,077.8446 ns |  52,037.143 ns |
|                 BisectToEnumerable |    50 |    100 |      30.221 ns |     0.3635 ns |     0.3222 ns |      30.290 ns |
|       BisectVerticallyToEnumerable |    50 |    100 |      28.996 ns |     0.4477 ns |     0.3969 ns |      29.063 ns |
|     BisectHorizontallyToEnumerable |    50 |    100 |      28.397 ns |     0.3681 ns |     0.3443 ns |      28.314 ns |
| BisectHorizontallyNativeEnumerable |    50 |    100 |      26.590 ns |     0.3960 ns |     0.3307 ns |      26.597 ns |
|                             Bisect |   100 |     10 |      12.218 ns |     0.1793 ns |     0.3936 ns |      12.124 ns |
|                   BisectVertically |   100 |     10 |      10.262 ns |     0.0399 ns |     0.0311 ns |      10.250 ns |
|                 BisectHorizontally |   100 |     10 |      10.500 ns |     0.0528 ns |     0.0441 ns |      10.480 ns |
|                    BisectRecursive |   100 |     10 |   5,310.233 ns |    64.3777 ns |    53.7583 ns |   5,316.383 ns |
|                 BisectToEnumerable |   100 |     10 |      31.569 ns |     0.5873 ns |     0.8609 ns |      31.280 ns |
|       BisectVerticallyToEnumerable |   100 |     10 |      29.221 ns |     0.6088 ns |     0.8535 ns |      29.010 ns |
|     BisectHorizontallyToEnumerable |   100 |     10 |      30.038 ns |     0.5824 ns |     1.2785 ns |      29.845 ns |
| BisectHorizontallyNativeEnumerable |   100 |     10 |      26.052 ns |     0.3928 ns |     0.3280 ns |      26.166 ns |
|                             Bisect |   100 |     50 |      11.746 ns |     0.0187 ns |     0.0146 ns |      11.748 ns |
|                   BisectVertically |   100 |     50 |      10.283 ns |     0.0505 ns |     0.0472 ns |      10.261 ns |
|                 BisectHorizontally |   100 |     50 |       9.890 ns |     0.0262 ns |     0.0233 ns |       9.893 ns |
|                    BisectRecursive |   100 |     50 |  51,831.080 ns |   236.9335 ns |   210.0354 ns |  51,763.559 ns |
|                 BisectToEnumerable |   100 |     50 |      30.176 ns |     0.2253 ns |     0.1997 ns |      30.203 ns |
|       BisectVerticallyToEnumerable |   100 |     50 |      28.489 ns |     0.4764 ns |     0.4223 ns |      28.588 ns |
|     BisectHorizontallyToEnumerable |   100 |     50 |      28.417 ns |     0.2325 ns |     0.1816 ns |      28.466 ns |
| BisectHorizontallyNativeEnumerable |   100 |     50 |      26.046 ns |     0.5336 ns |     0.4991 ns |      26.089 ns |
|                             Bisect |   100 |    100 |      11.512 ns |     0.0251 ns |     0.0223 ns |      11.511 ns |
|                   BisectVertically |   100 |    100 |       9.872 ns |     0.0263 ns |     0.0220 ns |       9.875 ns |
|                 BisectHorizontally |   100 |    100 |       9.877 ns |     0.0142 ns |     0.0119 ns |       9.880 ns |
|                    BisectRecursive |   100 |    100 | 111,830.037 ns |   866.6556 ns |   768.2677 ns | 112,161.945 ns |
|                 BisectToEnumerable |   100 |    100 |      31.524 ns |     0.6515 ns |     0.7503 ns |      31.514 ns |
|       BisectVerticallyToEnumerable |   100 |    100 |      28.466 ns |     0.2780 ns |     0.2465 ns |      28.503 ns |
|     BisectHorizontallyToEnumerable |   100 |    100 |      28.400 ns |     0.1862 ns |     0.1742 ns |      28.370 ns |
| BisectHorizontallyNativeEnumerable |   100 |    100 |      25.662 ns |     0.1236 ns |     0.1156 ns |      25.646 ns |

## Area Enumeration
Because `Area` implements `IEnumerable`, it would cause a lot of breaking changes to attempt to change the method of iteration used in the typical enumerator (and would be undesirable in a lot of circumstances either way).  Therefore, instead of changing the default enumerator, we instead provide a `FastEnumerator()` which provides a faster method to enumerate with a `foreach` loop.

However, `IReadOnlyArea` is an interface, and therefore much of the implementation is up to library consumers.  The default implementation uses a list and hash set paired together; however GoRogue, for example, provides an implementation called `MultiArea`, which takes multiple areas and exposes them as a single one, as well as `PolygonArea`, which represents polygons.  This makes it tough to determine what the "fastest" way to iterate is; for an area, it is certainly to access points by index using the area's indexer; but that is actually quite slow for GoRogue's `MultiArea` because the indexer is not nearly as efficient.

Therefore, this PR adds a flag `UseIndexEnumeration` to the area interface (with a default value of `false`, so as to avoid breaking backwards compatibility).  When this flag is set to true by implementations (such as Area), the enumerator provided will use indexing.  If it's set to false, it will default to the `IEnumerable` implementation of the interface itself.  This allows different implementations to specify proper iteration methods.


Here are the performance results of the new `FastEnumerator` function's enumerator:

|                                  Method | Size |            Mean |         Error |        StdDev |
|---------------------------------------- |----- |----------------:|--------------:|--------------:|
|                     BenchmarkEnumerable |   10 |       542.17 ns |      1.213 ns |      1.013 ns |
|           BenchmarkEnumerableAsConcrete |   10 |       401.20 ns |      2.106 ns |      1.970 ns |
|                 BenchmarkFastEnumerable |   10 |       267.44 ns |      0.887 ns |      0.692 ns |
|       BenchmarkFastEnumerableAsConcrete |   10 |       269.99 ns |      1.755 ns |      1.641 ns |
|                      BenchmarkManualFor |   10 |       362.09 ns |      3.069 ns |      2.871 ns |
|           BenchmarkManualForCachedCount |   10 |       196.27 ns |      0.795 ns |      0.705 ns |
|            BenchmarkManualForAsConcrete |   10 |        76.83 ns |      0.497 ns |      0.441 ns |
| BenchmarkManualForAsConcreteCachedCount |   10 |        76.36 ns |      0.369 ns |      0.345 ns |
|                     BenchmarkEnumerable |  100 |    52,136.04 ns |    373.036 ns |    330.686 ns |
|           BenchmarkEnumerableAsConcrete |  100 |    37,598.42 ns |    117.548 ns |    104.203 ns |
|                 BenchmarkFastEnumerable |  100 |    25,265.49 ns |    163.274 ns |    152.727 ns |
|       BenchmarkFastEnumerableAsConcrete |  100 |    25,128.46 ns |     68.585 ns |     60.798 ns |
|                      BenchmarkManualFor |  100 |    35,350.81 ns |     46.764 ns |     39.050 ns |
|           BenchmarkManualForCachedCount |  100 |    18,851.83 ns |     29.275 ns |     22.856 ns |
|            BenchmarkManualForAsConcrete |  100 |     7,127.25 ns |     10.012 ns |      8.360 ns |
| BenchmarkManualForAsConcreteCachedCount |  100 |     7,118.29 ns |     63.503 ns |     56.294 ns |
|                     BenchmarkEnumerable |  200 |   206,361.33 ns |    122.541 ns |    102.327 ns |
|           BenchmarkEnumerableAsConcrete |  200 |   150,827.81 ns |  1,164.635 ns |  1,089.400 ns |
|                 BenchmarkFastEnumerable |  200 |   100,393.31 ns |    239.670 ns |    200.135 ns |
|       BenchmarkFastEnumerableAsConcrete |  200 |   100,421.90 ns |    132.298 ns |    103.289 ns |
|                      BenchmarkManualFor |  200 |   140,725.61 ns |    221.211 ns |    172.707 ns |
|           BenchmarkManualForCachedCount |  200 |    75,417.81 ns |     82.344 ns |     72.996 ns |
|            BenchmarkManualForAsConcrete |  200 |    28,573.40 ns |     58.039 ns |     51.450 ns |
| BenchmarkManualForAsConcreteCachedCount |  200 |    28,299.78 ns |     74.319 ns |     65.882 ns |

As usual, the new enumerator is about 2x faster in some cases, and is roughly competitive with a for loop that doesn't cache the result of the `Count` property.  One notable exception here is the manual for loops do much better than anything else on a value of type `Area`, rather than `IReadOnlyArea`.  I believe this is because the compiler is able to see the underlying structure is a `List` in this case and apply optimizations.  In practice, I've found this difference in enumeration speed makes a difference only rarely, but nonetheless I was never able to make the `FastEnumerator` significantly slower than regular enumeration, so it seems to provide no downside.  I did find one test case involving GoRogue's `MultiArea` where there was a _slight_ decrease with `FastEnumerator()` however nothing significant enough to affect code in really any realistic case (and the benefits it creates with other types would seem to far outweigh that) In either case, this produces no breaking changes for users, since the initial iterator remains untouched.